Fix redundant casts

+++ ballerina-shell/modules/shell-core/src/main/java/io/ballerina/shell/utils/StringUtils.java
@@ -172,7 +172,7 @@ public class StringUtils {
      * @return Reformatted unicode string.
      */
     public static String convertUnicode(char character) {
-        return "\\u{" + Integer.toHexString((int) character) + "}";
+        return "\\u{" + Integer.toHexString(character) + "}";
     }
 
     /**
